# Switch to json-iterator for faster JSON encoding/decoding