--- a/cmd/restic/cmd_forget.go
+++ b/cmd/restic/cmd_forget.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"io"
 
 	"github.com/restic/restic/internal/errors"
+	"github.com/restic/restic/internal/json"
 	"github.com/restic/restic/internal/restic"
 	"github.com/spf13/cobra"
 )