--- a/cmd/restic/cmd_key.go
+++ b/cmd/restic/cmd_key.go
@@ -2,12 +2,12 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"io/ioutil"
 	"os"
 	"strings"
 
 	"github.com/restic/restic/internal/errors"
+	"github.com/restic/restic/internal/json"
 	"github.com/restic/restic/internal/repository"
 	"github.com/restic/restic/internal/restic"
 	"github.com/restic/restic/internal/ui/table"