--- a/internal/restic/blob_test.go
+++ b/internal/restic/blob_test.go
@@ -1,8 +1,9 @@
 package restic
 
 import (
-	"encoding/json"
 	"testing"
+
+	"github.com/restic/restic/internal/json"
 )
 
 var blobTypeJSON = []struct {