What is the purpose of this change? What does it change?
--------------------------------------------------------

All JSON encoding/decoding outside the tests is now done by [json-iterator](http://jsoniter.com) instead of standard Go encoding/json to boost performance.

internal/index benchmarks:

```
name         old time/op  new time/op  delta
IndexNew-8    175µs ± 2%   178µs ± 2%   +1.58%  (p=0.000 n=20+20)
IndexSave-8   4.02s ± 2%   3.04s ± 3%  -24.24%  (p=0.000 n=20+20)
```

internal/repository:

```
name              old time/op    new time/op    delta
DecodeIndex-8       13.7µs ± 3%     6.8µs ± 5%  -50.72%  (p=0.000 n=20+20)
LoadTree-8           431µs ± 6%     275µs ± 3%  -36.08%  (p=0.000 n=20+20)
LoadAndDecrypt-8    13.1ms ± 4%    13.9ms ± 2%   +5.53%  (p=0.000 n=20+20)
LoadIndex-8         32.8ms ± 4%    22.6ms ± 4%  -31.04%  (p=0.000 n=20+20)
```

backend/rest shows no significant changes.

Although jsoniter promises [100% compatibility](http://jsoniter.com/migrate-from-go-std.html), it doesn't quite live up to that promise, as errors are reported differently (it has no UnmarshalTypeError). The old index format detection code had to be changed to handle this.

Was the change discussed in an issue or in the forum before?
------------------------------------------------------------
In #838, @middelink demonstrated the performance issues with JSON and @fd0 suggested jsoniter (and some other libraries, which I didn't try). I'm not sure if it closes that issue.

Checklist
---------

- [x] I have read the [Contribution Guidelines](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#providing-patches)
- [x] I have enabled [maintainer edits for this PR](https://help.github.com/en/github/collaborating-with-issues-and-pull-requests/allowing-changes-to-a-pull-request-branch-created-from-a-fork)
- [x] I have added tests for all changes in this PR
- [x] I have run `gofmt` on the code in all commits
- [x] All commit messages are formatted in the same style as [the other commits in the repo](https://github.com/restic/restic/blob/master/CONTRIBUTING.md#git-commits)
- [x] I'm done, this Pull Request is ready for review