@buschjost 2020-06-12
I am running with this merged for over a month now and it is working fine for me.