@aawsome 2020-07-10
Seems like we are getting merge conflicts after #2709 was merged.

@greatroar would you mind rebasing this?

Also, this deserves a changelog entry IMO.