@greatroar 2020-07-10
Rebased. Also upgraded to json-iterator 1.1.10 and cleaned up the go.{mod,sum} files (the trick is to use `go get` *without* `-u`, I just learned). I did not also update the benchmark figures, but if new ones are desired I can provide them.