@MichaelEischer 2021-12-30
I've rebased the PR to the current master. From some initial performance tests it looks like the performance benefit is much smaller than before.