@MichaelEischer 2021-12-30
Results for `go test -bench 'Save|Decode|Load' ./internal/repository`

```
name                 old time/op    new time/op    delta
DecodeIndex             4.98s ± 3%     2.94s ± 4%  -40.88%  (p=0.000 n=10+10)
DecodeIndexParallel     5.07s ± 6%     2.78s ± 6%  -45.08%  (p=0.000 n=10+10)
SaveAndEncrypt         51.9ms ± 4%    52.6ms ± 9%     ~     (p=0.780 n=9+10)
LoadTree                367µs ± 8%     226µs ± 8%  -38.38%  (p=0.000 n=10+10)
LoadBlob               11.7ms ± 2%    11.8ms ± 5%     ~     (p=0.842 n=9+10)
LoadAndDecrypt         12.4ms ± 5%    12.5ms ± 7%     ~     (p=0.481 n=10+10)
LoadIndex              32.3ms ± 5%    22.8ms ± 5%  -29.53%  (p=0.000 n=10+9)

name                 old speed      new speed      delta
SaveAndEncrypt       80.2MB/s ± 7%  79.9MB/s ± 8%     ~     (p=1.000 n=10+10)
LoadBlob             85.5MB/s ± 2%  84.9MB/s ± 5%     ~     (p=0.842 n=9+10)
LoadAndDecrypt       80.9MB/s ± 5%  80.1MB/s ± 7%     ~     (p=0.481 n=10+10)

name                 old alloc/op   new alloc/op   delta
SaveAndEncrypt         27.2MB ± 0%    27.2MB ± 0%   +0.00%  (p=0.000 n=10+10)

name                 old allocs/op  new allocs/op  delta
SaveAndEncrypt            136 ± 0%       139 ± 0%   +2.21%  (p=0.000 n=10+10)
```

As `LoadIndex` includes all steps like loading and decrypting the index, this is the speed-up that could be expected in practice.