@greatroar 2022-02-06
Rebased again and updated the benchmark figures in the commit message.