@MichaelEischer 2022-04-30
I've run a few tests with this PR. For a 550GB repo it reduces the user CPU time for `check` from 645 to 450 seconds and the system time from 156 to 135 seconds. That's essentially the 30-40% speed improvements also shown by the benchmarks. The maximum memory usage increased from 2GB to 2.5GB (it's that high because the repository contains several 300MB tree blobs).

Using `GOGC=10` the peak memory usage has shrunk to 1.25GB and 1.4GB, respectively, while the user time increased to 892 and 758 seconds. As restic already requires quite a bit of memory, I'm not sure whether the change is worth it. Especially as other commands will benefit far less.