@MichaelEischer 2022-05-01
As the memory usage of restic is already quite high, I think trading a higher memory usage for a limited speed increase is probably not worth it at the moment. So I'm closing this PR, but anyways thanks for trying to improve restic!