@greatroar 2022-05-01
That's fair. I never checked the memory use of this.